collapse schema cache into client
