+++ pkg/schemaregistry/client.go
@@ -16,20 +16,39 @@ package schemaregistry
 
 import (
 	"context"
+	"fmt"
 
+	"github.com/conduitio/conduit-commons/rabin"
 	"github.com/conduitio/conduit/pkg/foundation/cerrors"
 	"github.com/conduitio/conduit/pkg/foundation/log"
-	"github.com/conduitio/conduit/pkg/schemaregistry/internal"
 	"github.com/twmb/franz-go/pkg/sr"
+	"github.com/twmb/go-cache/cache"
 )
 
 // Client is a schema registry client that caches schemas. It is safe for
-// concurrent use.
+// concurrent use. The client caches schemas by ID, their subject/fingerprint and
+// subject/version. Fingerprints are calculated using the Rabin algorithm.
 type Client struct {
 	logger log.CtxLogger
 	client sr.Client
 
-	cache internal.SchemaCache
+	idCache                 *cache.Cache[int, sr.Schema]
+	subjectFingerprintCache *cache.Cache[subjectFingerprint, sr.SubjectSchema]
+	subjectVersionCache     *cache.Cache[subjectVersion, sr.SubjectSchema]
+}
+
+type (
+	subjectVersion     string
+	subjectFingerprint string
+)
+
+func newSubjectVersion(subject string, version int) subjectVersion {
+	return subjectVersion(fmt.Sprintf("%s:%d", subject, version))
+}
+
+func newSubjectFingerprint(subject string, text string) subjectFingerprint {
+	fingerprint := rabin.Bytes([]byte(text))
+	return subjectFingerprint(fmt.Sprintf("%s:%d", subject, fingerprint))
 }
 
 var _ RegistryWithCheck = (*Client)(nil)
@@ -44,12 +63,16 @@ func NewClient(logger log.CtxLogger, opts ...sr.ClientOpt) (*Client, error) {
 
 	client, err := sr.NewClient(append(defaultOpts, opts...)...)
 	if err != nil {
-		return nil, err
+		return nil, cerrors.Errorf("failed to create schema registry client: %w", err)
 	}
 
 	return &Client{
 		logger: logger.WithComponent("schemaregistry.Client"),
 		client: *client,
+
+		idCache:                 cache.New[int, sr.Schema](),
+		subjectFingerprintCache: cache.New[subjectFingerprint, sr.SubjectSchema](),
+		subjectVersionCache:     cache.New[subjectVersion, sr.SubjectSchema](),
 	}, nil
 }
 
@@ -59,7 +82,9 @@ func NewClient(logger log.CtxLogger, opts ...sr.ClientOpt) (*Client, error) {
 // was successful.
 func (c *Client) CreateSchema(ctx context.Context, subject string, schema sr.Schema) (sr.SubjectSchema, error) {
 	logEvent := c.logger.Trace(ctx).Str("operation", "CreateSchema").Str("subject", subject)
-	ss, err := c.cache.GetBySubjectText(subject, schema.Schema, func() (sr.SubjectSchema, error) {
+
+	sfp := newSubjectFingerprint(subject, schema.Schema)
+	ss, err, _ := c.subjectFingerprintCache.Get(sfp, func() (sr.SubjectSchema, error) {
 		logEvent.Msg("schema cache miss")
 		logEvent = nil // disable output for hit
 
@@ -70,7 +95,7 @@ func (c *Client) CreateSchema(ctx context.Context, subject string, schema sr.Sch
 
 		ss, err := c.client.CreateSchema(ctx, subject, schema)
 		if err != nil {
-			return ss, err
+			return ss, cerrors.Errorf("failed to create schema with subject %q: %w", subject, err)
 		}
 
 		if !subjectExists {
@@ -83,10 +108,12 @@ func (c *Client) CreateSchema(ctx context.Context, subject string, schema sr.Sch
 				}
 			}
 		}
+		c.idCache.Set(ss.ID, ss.Schema)
+		c.subjectVersionCache.Set(newSubjectVersion(ss.Subject, ss.Version), ss)
 		return ss, nil
 	})
 	if err != nil {
-		return sr.SubjectSchema{}, cerrors.Errorf("failed to create schema with subject %q: %w", subject, err)
+		return sr.SubjectSchema{}, err
 	}
 	logEvent.Msg("schema cache hit")
 	return ss, nil
@@ -99,13 +126,18 @@ func (c *Client) CreateSchema(ctx context.Context, subject string, schema sr.Sch
 // cache will not have an effect on methods that return a sr.SubjectSchema.
 func (c *Client) SchemaByID(ctx context.Context, id int) (sr.Schema, error) {
 	logEvent := c.logger.Trace(ctx).Str("operation", "SchemaByID").Int("id", id)
-	s, err := c.cache.GetByID(id, func() (sr.Schema, error) {
+
+	s, err, _ := c.idCache.Get(id, func() (sr.Schema, error) {
 		logEvent.Msg("schema cache miss")
 		logEvent = nil // disable output for hit
-		return c.client.SchemaByID(ctx, id)
+		ss, err := c.client.SchemaByID(ctx, id)
+		if err != nil {
+			return sr.Schema{}, cerrors.Errorf("failed to get schema with ID %q: %w", id, err)
+		}
+		return ss, nil
 	})
 	if err != nil {
-		return sr.Schema{}, cerrors.Errorf("failed to get schema with ID %q: %w", id, err)
+		return sr.Schema{}, err
 	}
 	logEvent.Msg("schema cache hit")
 	return s, nil
@@ -118,13 +150,21 @@ func (c *Client) SchemaBySubjectVersion(ctx context.Context, subject string, ver
 	// TODO handle latest version separately, let caller define timeout after
 	//  which the latest cached version should be downloaded again from upstream
 	logEvent := c.logger.Trace(ctx).Str("operation", "SchemaBySubjectVersion").Str("subject", subject).Int("version", version)
-	ss, err := c.cache.GetBySubjectVersion(subject, version, func() (sr.SubjectSchema, error) {
+
+	sv := newSubjectVersion(subject, version)
+	ss, err, _ := c.subjectVersionCache.Get(sv, func() (sr.SubjectSchema, error) {
 		logEvent.Msg("schema cache miss")
 		logEvent = nil // disable output for hit
-		return c.client.SchemaByVersion(ctx, subject, version)
+		ss, err := c.client.SchemaByVersion(ctx, subject, version)
+		if err != nil {
+			return ss, cerrors.Errorf("failed to get schema with subject %q and version %q: %w", subject, version, err)
+		}
+		c.idCache.Set(ss.ID, ss.Schema)
+		c.subjectFingerprintCache.Set(newSubjectFingerprint(ss.Subject, ss.Schema.Schema), ss)
+		return ss, nil
 	})
 	if err != nil {
-		return sr.SubjectSchema{}, cerrors.Errorf("failed to get schema with subject %q and version %q: %w", subject, version, err)
+		return sr.SubjectSchema{}, err
 	}
 	logEvent.Msg("schema cache hit")
 	return ss, nil
